fset-2504 wip only 1 failing IT now

+++ app/services/assessor/AssessorService.scala
@@ -234,7 +234,7 @@ trait AssessorService {
     }
   }
 
-  def countSubmittedAvailability(): Future[Int] = {
+  def countSubmittedAvailability(): Future[Long] = {
     assessorRepository.countSubmittedAvailability
   }
